Stats: move Post Trends to /stats
